# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "SecureAuth MFA OTP & CAPTCHA API",
        "description": "Complete API documentation for the SecureAuth application.\n\n## Running modes\n\n### Monolith mode (`npm run start`)\n- Single server on `:3000` implements registration (OTP → CAPTCHA → TOTP setup) and login (password → TOTP).\n\n### Multi-server mode (`npm run start:multi`)\n- **OTP/UI** server on `:3000`\n- **CAPTCHA** server on `:5000`\n- **MFA** server on `:8000`\n\nIn multi-server mode, login is **password → login-OTP → CAPTCHA → MFA**.\n\n## Flows\n\n### Registration (monolith & multi-server OTP server)\n1. `POST /api/register` — Register & send OTP (email or SMS)\n2. `POST /api/verify-email` — Verify the 6-digit OTP\n3. `GET /api/captcha` — Get SVG CAPTCHA challenge (registration)\n4. `POST /api/verify-captcha` — Verify CAPTCHA & get TOTP QR code\n5. Scan QR code in Google Authenticator\n\n### Login (multi-server)\n1. `POST /api/login` — Password check, then send login OTP\n2. `POST /api/verify-login-otp` — Verify login OTP\n3. `GET /api/captcha` — Get SVG CAPTCHA challenge (login CAPTCHA server)\n4. `POST /api/verify-captcha` — Verify CAPTCHA (login CAPTCHA server)\n5. `POST /api/verify-mfa` — Verify TOTP (MFA server)\n\n### Session\n- `GET /api/me` — Get current logged-in user\n- `POST /api/logout` — Destroy session",
        "version": "1.1.0",
        "contact": {
            "name": "SecureAuth",
...
        }
    },
    "servers": [
        {
...
        {
...
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Registration"
                ],
                "summary": "Register user & send OTP",
                "description": "Register a new user and send a 6-digit OTP to their email address or phone number via SMS.\n\n**DEV MODE:** If `EMAIL_USER`/`EMAIL_PASS` are not set in `.env`, the OTP is logged to the server console instead of being emailed. Similarly for Twilio SMS.",
                "operationId": "registerUser",
                "servers": [
                    {
                        "url": "http://localhost:{otpPort}",
                        "description": "OTP/UI server only",
                        "variables": {
                            "otpPort": {
                                "default": "3000",
                                "description": "Your OTP/UI port (e.g. 3500)."
                            }
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            },
                            "examples": {
                                "email": {
                                    "summary": "Register with Email",
                                    "value": {
                                        "name": "John Doe",
                                        "email": "john@example.com",
                                        "phone": "",
                                        "password": "SecurePass123!",
                                        "method": "email"
                                    }
                                },
                                "phone": {
                                    "summary": "Register with Phone (SMS)",
                                    "value": {
                                        "name": "Jane Doe",
                                        "email": "",
                                        "phone": "+1234567890",
                                        "password": "SecurePass456!",
                                        "method": "phone"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegisterResponse"
                                },
                                "example": {
                                    "success": true,
                                    "identifier": "john@example.com",
                                    "message": "Verification email sent."
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Email or phone already registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "error": "Email already registered."
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to send email/SMS",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "error": "Failed to send email after multiple attempts. Check configuration."
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/verify-email": {
            "post": {
                "tags": [
                    "Registration"
                ],
                "summary": "Verify OTP code",
                "description": "Verify the 6-digit OTP code that was sent during registration. On success, the next step is to complete the CAPTCHA challenge.",
                "operationId": "verifyEmailOtp",
                "servers": [
                    {
                        "url": "http://localhost:{otpPort}",
                        "description": "OTP/UI server only",
                        "variables": {
                            "otpPort": {
                                "default": "3000",
                                "description": "Your OTP/UI port (e.g. 3500)."
                            }
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyOtpRequest"
                            },
                            "example": {
                                "identifier": "john@example.com",
                                "otp": "482931"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP verified — CAPTCHA required next",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerifyOtpResponse"
                                },
                                "example": {
                                    "success": true,
                                    "requireCaptcha": true,
                                    "message": "OTP verified. Please complete the CAPTCHA."
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid or expired OTP",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "error": "Invalid or expired OTP."
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/captcha": {
            "get": {
                "tags": [
                    "Registration"
                ],
                "summary": "Get CAPTCHA challenge",
                "description": "Generate and return a new SVG CAPTCHA image.\n\n- On `http://localhost:3000` this is the **registration CAPTCHA**.\n- On `http://localhost:5000` this is the **login CAPTCHA** (requires login-OTP step first).\n\nThe CAPTCHA text is stored in the server session. Each call generates a new CAPTCHA (the old one is replaced).\n\n**CAPTCHA Config:** 5 characters, 3 noise lines, colored.",
                "operationId": "getCaptcha",
                "servers": [
                    {
                        "url": "http://localhost:3000",
                        "description": "Registration CAPTCHA (OTP/UI server)"
                    },
                    {
                        "url": "http://localhost:5000",
                        "description": "Login CAPTCHA (CAPTCHA service)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CAPTCHA image generated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CaptchaResponse"
                                },
                                "example": {
                                    "success": true,
                                    "captchaSvg": "<svg xmlns='http://www.w3.org/2000/svg' width='150' height='50'>...</svg>"
                                }
                            }
                        }
                    }
                    ,
                    "401": {
                        "description": "Login CAPTCHA prerequisites not satisfied (multi-server CAPTCHA service only)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "error": "OTP step not completed yet."
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/verify-captcha": {
            "post": {
                "tags": [
                    "Registration"
                ],
                "summary": "Verify CAPTCHA & complete registration",
                "description": "Verify a CAPTCHA answer.\n\n- On `http://localhost:3000` this completes **registration** and returns a TOTP QR code.\n- On `http://localhost:5000` this completes the **login CAPTCHA step** and returns the next URL (MFA).\n\n**Prerequisites:**\n- CAPTCHA must have been fetched via `GET /api/captcha`\n- For registration (OTP/UI server): OTP must already be verified\n- For login (CAPTCHA service): login OTP must already be verified",
                "operationId": "verifyCaptcha",
                "servers": [
                    {
                        "url": "http://localhost:3000",
                        "description": "Registration CAPTCHA verify (OTP/UI server)"
                    },
                    {
                        "url": "http://localhost:5000",
                        "description": "Login CAPTCHA verify (CAPTCHA service)"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyCaptchaRequest"
                            },
                            "example": {
                                "identifier": "john@example.com",
                                "captchaAnswer": "AbCdE"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CAPTCHA verified (response differs by server/mode)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/VerifyCaptchaResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/LoginCaptchaVerifiedResponse"
                                        }
                                    ]
                                },
                                "examples": {
                                    "registration": {
                                        "summary": "Registration completion (OTP/UI server)",
                                        "value": {
                                            "success": true,
                                            "message": "CAPTCHA verified! Please scan the QR code to setup your Authenticator app.",
                                            "qrCodeUrl": "data:image/png;base64,iVBORw0KGgo...",
                                            "secret": "JBSWY3DPEHPK3PXP"
                                        }
                                    },
                                    "loginCaptcha": {
                                        "summary": "Login CAPTCHA completion (CAPTCHA service)",
                                        "value": {
                                            "success": true,
                                            "requireMfa": true,
                                            "mfaUrl": "http://localhost:8000/",
                                            "message": "CAPTCHA verified. Continue with MFA."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "OTP not verified or incorrect CAPTCHA",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "examples": {
                                    "otpNotVerified": {
                                        "summary": "OTP not verified yet",
                                        "value": {
                                            "error": "OTP not verified yet."
                                        }
                                    },
                                    "wrongCaptcha": {
                                        "summary": "Incorrect CAPTCHA",
                                        "value": {
                                            "error": "Incorrect CAPTCHA. Please try again."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "QR code generation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "error": "Error generating QR code."
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "RegisterRequest": {
                "type": "object",
                "required": [
                    "name",
                    "password",
                    "method"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "User's full name",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "description": "Email address (required when method = 'email')",
                        "example": "john@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "description": "Phone number with country code (required when method = 'phone')",
                        "example": "+1234567890"
                    },
                    "password": {
                        "type": "string",
                        "description": "User password",
                        "example": "SecurePass123!"
                    },
                    "method": {
                        "type": "string",
                        "enum": [
                            "email",
                            "phone"
                        ],
                        "description": "Verification method — 'email' or 'phone'",
                        "example": "email"
                    }
                }
            },
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "description": "Error message",
                        "example": "Something went wrong."
                    }
                }
            }
        }
    }
}